+++ webview/src/android_webview/BUILD.gn
@@ -729,24 +729,10 @@ android_library("browser_java") {
     "java/src/org/chromium/base/WeakContext.java",
     "java/src/org/chromium/components/navigation_interception/NavigationParams.java",
     "java/src/org/chromium/content/R.java",
-    "java/src/org/chromium/content/app/ChildProcessService.java",
     "java/src/org/chromium/content/app/LibraryLoader.java",
     "java/src/org/chromium/content/app/NativeLibraries.java",
-    "java/src/org/chromium/content/app/PrivilegedProcessService.java",
-    "java/src/org/chromium/content/app/PrivilegedProcessService0.java",
-    "java/src/org/chromium/content/app/PrivilegedProcessService1.java",
-    "java/src/org/chromium/content/app/PrivilegedProcessService2.java",
-    "java/src/org/chromium/content/app/SandboxedProcessService.java",
-    "java/src/org/chromium/content/app/SandboxedProcessService0.java",
-    "java/src/org/chromium/content/app/SandboxedProcessService1.java",
-    "java/src/org/chromium/content/app/SandboxedProcessService2.java",
-    "java/src/org/chromium/content/app/SandboxedProcessService3.java",
-    "java/src/org/chromium/content/app/SandboxedProcessService4.java",
-    "java/src/org/chromium/content/app/SandboxedProcessService5.java",
     "java/src/org/chromium/content/browser/ActivityContentVideoViewDelegate.java",
     "java/src/org/chromium/content/browser/AndroidBrowserProcess.java",
-    "java/src/org/chromium/content/browser/ChildProcessConnection.java",
-    "java/src/org/chromium/content/browser/ChildProcessLauncher.java",
     "java/src/org/chromium/content/browser/ContentSettings.java",
     "java/src/org/chromium/content/browser/ContentVideoView.java",
     "java/src/org/chromium/content/browser/ContentVideoViewContextDelegate.java",
@@ -763,7 +749,6 @@ android_library("browser_java") {
     "java/src/org/chromium/content/browser/FileDescriptorInfo.java",
     "java/src/org/chromium/content/browser/HeapStatsLogger.java",
     "java/src/org/chromium/content/browser/InterstitialPageDelegateAndroid.java",
-    "java/src/org/chromium/content/browser/JavascriptInterface.java",
     "java/src/org/chromium/content/browser/JellyBeanContentView.java",
     "java/src/org/chromium/content/browser/LoadUrlParams.java",
     "java/src/org/chromium/content/browser/LocationProvider.java",
@@ -791,7 +776,6 @@ android_library("browser_java") {
     "java/src/org/chromium/content/browser/accessibility/JellyBeanAccessibilityInjector.java",
     "java/src/org/chromium/content/browser/input/AdapterInputConnection.java",
     "java/src/org/chromium/content/browser/input/CursorController.java",
-    "java/src/org/chromium/content/browser/input/DateTimeChooserAndroid.java",
     "java/src/org/chromium/content/browser/input/DateTimePickerDialog.java",
     "java/src/org/chromium/content/browser/input/HandleView.java",
     "java/src/org/chromium/content/browser/input/ImeAdapter.java",
@@ -800,7 +784,6 @@ android_library("browser_java") {
     "java/src/org/chromium/content/browser/input/InsertionHandleController.java",
     "java/src/org/chromium/content/browser/input/MonthPicker.java",
     "java/src/org/chromium/content/browser/input/MonthPickerDialog.java",
-    "java/src/org/chromium/content/browser/input/SelectPopupDialog.java",
     "java/src/org/chromium/content/browser/input/SelectionHandleController.java",
     "java/src/org/chromium/content/browser/third_party/GestureDetector.java",
     "java/src/org/chromium/content/common/CleanupReference.java",
@@ -810,30 +793,17 @@ android_library("browser_java") {
     "java/src/org/chromium/content/common/ProcessInitException.java",
     "java/src/org/chromium/content/common/ResultCodes.java",
     "java/src/org/chromium/content/common/TraceEvent.java",
-    "java/src/org/chromium/media/AudioManagerAndroid.java",
-    "java/src/org/chromium/media/MediaCodecBridge.java",
-    "java/src/org/chromium/media/MediaPlayerBridge.java",
-    "java/src/org/chromium/media/MediaPlayerListener.java",
-    "java/src/org/chromium/media/VideoCapture.java",
     "java/src/org/chromium/media/WebAudioMediaCodecBridge.java",
-    "java/src/org/chromium/net/AndroidKeyStore.java",
-    "java/src/org/chromium/net/AndroidNetworkLibrary.java",
     "java/src/org/chromium/net/CertVerifyResultAndroid.java",
     "java/src/org/chromium/net/CertificateMimeType.java",
-    "java/src/org/chromium/net/GURLUtils.java",
-    "java/src/org/chromium/net/PrivateKeyType.java",
-    "java/src/org/chromium/net/ProxyChangeListener.java",
-    "java/src/org/chromium/net/X509Util.java",
     "java/src/org/chromium/ui/Clipboard.java",
     "java/src/org/chromium/ui/ColorPickerDialog.java",
     "java/src/org/chromium/ui/LocalizationUtils.java",
     "java/src/org/chromium/ui/R.java",
     "java/src/org/chromium/ui/SelectFileDialog.java",
-    "java/src/org/chromium/ui/UiUtils.java",
     "java/src/org/chromium/ui/ViewAndroid.java",
     "java/src/org/chromium/ui/ViewAndroidDelegate.java",
     "java/src/org/chromium/ui/WindowAndroid.java",
-    "java/src/org/chromium/ui/gfx/BitmapHelper.java",
     "java/src/org/chromium/ui/gfx/DeviceDisplayInfo.java",
     "java/src/org/chromium/ui/gfx/SurfaceTextureListener.java",
   ]
@@ -849,6 +819,7 @@ android_library("browser_java") {
     "//android_webview/proto:metrics_bridge_records_proto_java",
     "//base:base_java",
     "//base:content_uri_utils_java",
+    "//base:process_launcher_java",
     "//base/version_info/android:version_constants_java",
     "//build/android:build_java",
     "//cc/mojom:mojom_java",
